section the region into components
